ai-service/dataset/requirements.txt
requests==2.31.0
pandas>=1.3.0
scikit-learn==1.4.0
joblib==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0
